Adding my article Write Better + Some improvements in README.md
Hello,

I've discovered Markdown since two months or so, and it's such a pity
that it's not more well-known.

But that's not terribly strange either. It's hard to find the right
information about it. Even the wikipedia article needs a lot of
rewritings.

So first I wrote an article about why we must kill the idea that you
need Microsoft Word to write words. And why things like BBcode must die.

Then I wanted to launch a website with even more information.

Finally, I thought that maybe someone had the same idea.

And here I am :-)